gethostbyname

+++ c/container_of/main.c
@@ -1,6 +1,8 @@
 #include <stdio.h>
+#include <stdlib.h>
+#include <stdint.h>
 
-#define offsetof(TYPE, MEMBER) ((int)(&((TYPE *)0)->MEMBER))
+#define offsetof(TYPE, MEMBER) ((intptr_t)(&((TYPE *)0)->MEMBER))
 
 #define container_of(ptr, type, member) ({   \
 		const typeof( ((type *)0)->member ) *__mptr = (ptr); \
